chore: code refactoring
