+++ lib/enums/hatch_mark_alignment_enum.dart
@@ -1,0 +1,4 @@
+enum FlutterSliderHatchMarkAlignment {
+  left,
+  right,
+}
