+++ lib/enums/tooltip_direction_enum.dart
@@ -1,0 +1,5 @@
+enum FlutterSliderTooltipDirection {
+  top,
+  left,
+  right,
+}
