+++ lib/models/hatch_mark_label.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class FlutterSliderHatchMarkLabel {
+  final double? percent;
+  final Widget? label;
+
+  FlutterSliderHatchMarkLabel({
+    this.percent,
+    this.label,
+  }) : assert((label == null && percent == null) || (label != null && percent != null && percent >= 0));
+
+  @override
+  String toString() {
+    return '$percent-$label';
+  }
+}
