+++ lib/models/ignore_steps.dart
@@ -1,0 +1,11 @@
+class FlutterSliderIgnoreSteps {
+  final double? from;
+  final double? to;
+
+  FlutterSliderIgnoreSteps({this.from, this.to}) : assert(from != null && to != null && from <= to);
+
+  @override
+  String toString() {
+    return '$from-$to';
+  }
+}
