+++ lib/models/tooltip/tooltip_box.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class FlutterSliderTooltipBox {
+  final BoxDecoration? decoration;
+  final BoxDecoration? foregroundDecoration;
+  final Matrix4? transform;
+
+  const FlutterSliderTooltipBox({this.decoration, this.foregroundDecoration, this.transform});
+
+  @override
+  String toString() {
+    return '$decoration-$foregroundDecoration-$transform';
+  }
+}
