+++ lib/models/tooltip/tooltip_position_offset.dart
@@ -1,0 +1,13 @@
+class FlutterSliderTooltipPositionOffset {
+  double? top;
+  double? left;
+  double? right;
+  double? bottom;
+
+  FlutterSliderTooltipPositionOffset({this.top, this.left, this.right, this.bottom});
+
+  @override
+  String toString() {
+    return '$top-$left-$bottom-$right';
+  }
+}
