+++ lib/models/trackbar.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class FlutterSliderTrackBar {
+  final BoxDecoration? inactiveTrackBar;
+  final BoxDecoration? inactiveRightTrackBar;
+  final BoxDecoration? activeTrackBar;
+  final Color activeDisabledTrackBarColor;
+  final Color inactiveDisabledTrackBarColor;
+  final double activeTrackBarHeight;
+  final double inactiveTrackBarHeight;
+  final Widget? centralWidget;
+  final bool activeTrackBarDraggable;
+
+  const FlutterSliderTrackBar({
+    this.inactiveTrackBar,
+    this.activeTrackBar,
+    this.inactiveRightTrackBar,
+    this.activeDisabledTrackBarColor = const Color(0xffb5b5b5),
+    this.inactiveDisabledTrackBarColor = const Color(0xffe5e5e5),
+    this.activeTrackBarHeight = 3.5,
+    this.inactiveTrackBarHeight = 3,
+    this.centralWidget,
+    this.activeTrackBarDraggable = true,
+  }) : assert(activeTrackBarHeight > 0 && inactiveTrackBarHeight > 0);
+
+  @override
+  String toString() {
+    return '$inactiveTrackBar-$activeTrackBar-$activeDisabledTrackBarColor-$inactiveDisabledTrackBarColor-$activeTrackBarHeight-$inactiveTrackBarHeight-$centralWidget';
+  }
+}
