+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: another_xlider
 description: (Another Xlider) A material design slider and range slider, horizontal and vertical, with RTL support and lots of options and customizations for flutter
-version: 1.1.2
+version: 3.0.0
 homepage: https://github.com/loonix/another_xlider
 
 environment:
